Chapter 4: Add pixelsperscanline property for framebuffer node
pixelsperscanline defines the number of pixel elements per video memory line.
For performance reasons, or due to hardware restrictions, scan lines may be
padded to an amount of memory alignment. These padding pixel elements are
outside the area covered by HorizontalResolution and are not visible.

For direct frame buffer access, this number is used as a span between starts
of pixel lines in video memory. Based on the size of an individual pixel
element and PixelsPerScanline , the offset in video memory from
pixel element (x, y) to pixel element (x, y+1) has to be calculated
as “sizeof( PixelElement ) * PixelsPerScanLine”,
not “sizeof( PixelElement ) * HorizontalResolution”,
though in many cases those values can coincide.

This value can depend on video hardware and mode resolution.
GOP implementation is responsible for providing accurate value for this field.

Signed-off-by: Ajan Zhong <ajan.zhong@newfw.com>